DiffusionFlux

Documentation for DiffusionFlux.

DiffusionFlux.Properties

DiffusionFlux.D_Kn

DiffusionFlux.KozeneyCarman

DiffusionFlux.dgm_objects!

DiffusionFlux.flux_dgm!

DiffusionFlux.flux_ficks!

DiffusionFlux.interface_flux!

DiffusionFlux.Properties — Type

A structure to define the porous medis properties

ϵ : porosity

τ : tortuosity

pore_dia : pore diameter

part_dia : particle diameter

DiffusionFlux.D_Kn — Method

Function to calcuate the Knudsen diffusion coeffcients

Usage

D_Kn(thermo_obj, pm::Properties, T)

thermo_obj : SpeciesThermProperties (Refer IdealGas)

pm : Struct of the type Properties

T : Temperature (K)

DiffusionFlux.KozeneyCarman — Method

Function to calculate the permeability using Kozeney Carman relationship

Usage

KozeneyCarman(pm::Properties)

pm: Struct of the type Properties

DiffusionFlux.dgm_objects! — Method

Function to calculate binary diffusion coeffcients and Knudsen diffusion coeffcients, which are required for the evaluation of DGM fluxes

Usage

dgmobjects!(dgmobj::DGMObjects, pm::Properties, sp_trd, p::Float64, T::Float64 ,molwts::Array{Float64})

dgm_obj : A struct of the type DGMObjects which stores the diffusion coeffcients matrix

pm : Porous media properties (struct of the type Properties)

sptrd : Array of Species transport data, which is obtained by calling createtransport_data of TransportProperties

p : pressure in (Pa)

T : Temperature (K)

molwts : molecular weights vector

DiffusionFlux.flux_dgm! — Method

Function to calculate the mass fluxes (kg/m^2-s) at the interface between two cells using DGM The fluxes at the boundary cells are not evaluated in this function. If There are n cells, there will be n-1 internal cell faces

Usage

fluxdgm!(jks::Array{Array{T,1},1}, C::Array{Array{T,1},1}, pm::Properties, dgmobj::DGMObjects , molwts::Array{T}, Temp::T, δ::T)

jks : vector{vector} for storing the fluxes jks[n] stands for the flux at the interface between n and (n+1)th cell

C : vector{vector} concentrations in all cells

pm : Porous media properties (struct of the type Properties)

dgm_obj : A struct of the type DGMObjects which stores the diffusion coeffcients matrix

molwts : vector of molecular weights

DiffusionFlux.flux_ficks! — Method

Function to calculate the diffusion fluxes in Kg/m^2-s using Ficks law. The fluxes are corrected to ensure that the sum is zero

Usage

flux_ficks!(jks::Array{T}, Dkm::Array{T} ,C1::Array{T}, C2::Array{T}, δ::T)

jks : vector of fluxes

C1 : concentration vector

C2 : concentration vector

δ : distance between the cell centers

DiffusionFlux.interface_flux! — Method

Function to calcuate the mass flux (kg/m^2-s) at the inteface between flow channel and porous media. The pressure driven flux is assumed to be negligible at the inteface

Usage

interfaceflux!(jks::Array{T,1},Cch, Cpm, dgmobj::DGMObjects, molwts::Array{T}, Temp::T, δ::T)

jks : Storage for species flux (1D Array)

C_ch : concentration in the flow channel

C_pm : concentration in the porous media

dgm_obj : Diffusion matrix for DGM flux calculations

molwts : vector of molecular weights

Temp : Temperature (K)

δ : distance between the cell centers